rename util and a few test names
